requirements.txt
opencv-python
numpy
pillow
pyautogui
pytesseract